Adding documentation for examples
