+++ examples/src/main/java/io/dapr/examples/bindings/InputBindingController.java
@@ -1,0 +1,25 @@
+/*
+ * Copyright (c) Microsoft Corporation.
+ * Licensed under the MIT License.
+ */
+
+package io.dapr.examples.bindings;
+
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+import reactor.core.publisher.Mono;
+
+/**
+ * SpringBoot Controller to handle input binding.
+ */
+@RestController
+public class InputBindingController {
+
+  @PostMapping(path = "/bindingSample")
+  public Mono<Void> handleInputBinding(@RequestBody(required = false) byte[] body) {
+    return Mono.fromRunnable(() ->
+            System.out.println("Received message through binding: " + (body == null ? "" : new String(body))));
+  }
+
+}
