+++ examples/src/main/java/io/dapr/examples/bindings/docker-compose-single-kafka.yml
@@ -1,0 +1,14 @@
+version: '2'
+services:
+  zookeeper:
+    image: wurstmeister/zookeeper:latest
+    ports:
+      - "2181:2181"
+  kafka:
+    image: wurstmeister/kafka:latest
+    ports:
+      - "9092:9092"
+    environment:
+      KAFKA_ADVERTISED_HOST_NAME: 127.0.0.1
+      KAFKA_CREATE_TOPICS: "sample:1:1"
+      KAFKA_ZOOKEEPER_CONNECT: zookeeper:2181
