+++ examples/src/main/java/io/dapr/examples/pubsub/http/Publisher.java
@@ -16,18 +16,21 @@ import java.util.Collections;
  * 1. Build and install jars:
  * mvn clean install
  * 2. Run the program:
- * dapr run --app-id publisher --port 3006 -- mvn exec:java -pl=examples -Dexec.mainClass=io.dapr.examples.pubsub.http.Publisher
+ * dapr run --app-id publisher --port 3006 -- mvn exec:java -pl=examples -D exec.mainClass=io.dapr.examples.pubsub.http.Publisher
  */
 public class Publisher {
 
+  //Number of messages to be sent: 10
   private static final int NUM_MESSAGES = 10;
-
-  private static final String TOPIC_NAME = "message";
+  //The title of the topic to be used for publishing
+  private static final String TOPIC_NAME = "testingtopic";
 
   public static void main(String[] args) throws Exception {
+    //Creating the DaprClient: Using the default builder client produces an HTTP Dapr Client
     DaprClient client = new DaprClientBuilder(new DefaultObjectSerializer(), new DefaultObjectSerializer()).build();
     for (int i = 0; i < NUM_MESSAGES; i++) {
       String message = String.format("This is message #%d", i);
+      //Publishing messages
       client.publishEvent(TOPIC_NAME, message).block();
       System.out.println("Published message: " + message);
 
@@ -40,10 +43,11 @@ public class Publisher {
       }
     }
 
+    //Publishing a single bite: Example of non-string based content published
     client.publishEvent(
-        TOPIC_NAME,
-        new byte[] { 1 },
-        Collections.singletonMap("content-type", "application/octet-stream")).block();
+            TOPIC_NAME,
+            new byte[] { 1 },
+            Collections.singletonMap("content-type", "application/octet-stream")).block();
     System.out.println("Published one byte.");
 
     System.out.println("Done.");
