+++ examples/src/main/java/io/dapr/examples/pubsub/http/Subscriber.java
@@ -10,7 +10,6 @@ import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.CommandLineParser;
 import org.apache.commons.cli.DefaultParser;
 import org.apache.commons.cli.Options;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 /**
  * Service for subscriber.
@@ -19,7 +18,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
  * 2. Run the server:
  * dapr run --app-id subscriber --app-port 3000 --port 3005 -- mvn exec:java -pl=examples -Dexec.mainClass=io.dapr.examples.pubsub.http.Subscriber -Dexec.args="-p 3000"
  */
-@SpringBootApplication
 public class Subscriber {
 
   public static void main(String[] args) throws Exception {
